Split Lanbe into a VoictentLanbe and a VoictentItemLanbe
That is we can now act on a collection as a whole and not
just each item of a collection if needed. The mechanism
that streams a collection as a whole only triggers when the collection
has received an item on the previous tick, but did not receive an item\
on this tick of the engine. A "tick" is one loop of the platomities